Update drag for list (#35)
* Fix drag for list

* Fix drag for list

* Remove delay for insert animation

* Clean up
