+++ lib/src/model/motion_data.dart
@@ -3,23 +3,21 @@ import 'package:flutter/cupertino.dart';
 class MotionData {
   final Offset startOffset;
   final Offset endOffset;
-  final Duration duration;
 
   MotionData(
       {this.startOffset = Offset.zero,
       this.endOffset = Offset.zero,
-      this.duration = const Duration(milliseconds: 300)});
+      });
 
   MotionData copyWith({
     Offset? startOffset,
     Offset? endOffset,
-    Duration? duration,
     int? index,
   }) {
     return MotionData(
         startOffset: startOffset ?? this.startOffset,
         endOffset: endOffset ?? this.endOffset,
-        duration: duration ?? this.duration);
+        );
   }
 
   @override
